# db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package db

import (
	"database/sql"
)

type User struct {
	ID             int64          `db:"id" json:"id"`
	Email          string         `db:"email" json:"email"`
	Age            int64          `db:"age" json:"age"`
	Name           sql.NullString `db:"name" json:"name"`
	Bio            sql.NullString `db:"bio" json:"bio"`
	IsActive       bool           `db:"is_active" json:"is_active"`
	AdditionalData sql.NullString `db:"additional_data" json:"additional_data"`
	CreatedAt      sql.NullTime   `db:"created_at" json:"created_at"`
	UpdatedAt      sql.NullTime   `db:"updated_at" json:"updated_at"`
}