# generated/types.go
// Package generated provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.16.3 DO NOT EDIT.
package generated

import (
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	// Error Error message
	Error string `json:"error"`
}

// User defines model for User.
type User struct {
	// Age User age
	Age int `json:"age"`

	// Bio User biography (optional)
	Bio *string `json:"bio,omitempty"`

	// Email User email address
	Email openapi_types.Email `json:"email"`

	// Id User ID
	Id int64 `json:"id"`

	// IsActive Whether user is active (optional)
	IsActive *bool `json:"is_active,omitempty"`

	// Name User name (optional)
	Name *string `json:"name,omitempty"`
}

// UserRequest defines model for UserRequest.
type UserRequest struct {
	// Age User age
	Age int `json:"age"`

	// Bio User biography (optional)
	Bio *string `json:"bio,omitempty"`

	// Email User email address
	Email openapi_types.Email `json:"email"`

	// IsActive Whether user is active (optional)
	IsActive *bool `json:"is_active,omitempty"`

	// Name User name (optional)
	Name *string `json:"name,omitempty"`
}

// CreateUserJSONRequestBody defines body for CreateUser for application/json ContentType.
type CreateUserJSONRequestBody = UserRequest