# openapi.yaml
openapi: 3.0.3
info:
  title: User API
  description: User management API with validation
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local server
paths:
  /users:
    post:
      summary: Create a new user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{id}:
    get:
      summary: Get user by ID
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    User:
      type: object
      required:
        - id
        - email
        - age
      additionalProperties: false
      properties:
        id:
          type: integer
          format: int64
          description: User ID
        email:
          type: string
          format: email
          description: User email address
        age:
          type: integer
          minimum: 0
          description: User age
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: User name (optional)
        bio:
          type: string
          maxLength: 500
          description: User biography (optional)
        is_active:
          type: boolean
          default: true
          description: Whether user is active (optional)
    UserRequest:
      type: object
      required:
        - email
        - age
      additionalProperties: false
      properties:
        email:
          type: string
          format: email
          description: User email address
        age:
          type: integer
          minimum: 0
          description: User age
        name:
          type: string
          minLength: 1
          maxLength: 100
          description: User name (optional)
        bio:
          type: string
          maxLength: 500
          description: User biography (optional)
        is_active:
          type: boolean
          default: true
          description: Whether user is active (optional)
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message